split-utils * Move array-functions to utils/array.ts
