+++ src/containers/Admin/EditTab/ScooterPanel/ScooterPanel.tsx
@@ -3,9 +3,9 @@ import { useQuery } from '@apollo/client'
 import { Fieldset } from '@entur/form'
 import { FilterChip } from '@entur/chip'
 import { Operator } from '@entur/sdk/lib/mobility/types'
-import { toggleValueInList } from '../../../../utils'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { ALL_ACTIVE_OPERATOR_IDS } from '../../../../constants'
+import { toggleValueInList } from '../../../../utils/array'
 import './ScooterPanel.scss'
 import ScooterPanelQuery from './ScooterPanelQuery.mobility.graphql'
