+++ src/containers/Admin/EditTab/StopPlacePanel/PanelRow/Panelrow.tsx
@@ -3,9 +3,10 @@ import { Checkbox, TravelSwitch } from '@entur/form'
 import type { TravelSwitchProps } from '@entur/form'
 import { ExpandablePanel } from '@entur/expand'
 import { TransportMode } from '@entur/sdk'
-import { unique, isTransport } from '../../../../../utils'
+import { isTransport } from '../../../../../utils'
 import { Settings } from '../../../../../settings'
 import { StopPlaceWithLines } from '../../../../../types'
+import { unique } from '../../../../../utils/array'
 
 const PanelRow = ({
     onToggleStop,
