+++ src/containers/Admin/EditTab/StopPlacePanel/StopPlacePanel.tsx
@@ -3,7 +3,8 @@ import { Checkbox } from '@entur/form'
 import { Paragraph } from '@entur/typography'
 import { TransportMode } from '@entur/sdk'
 import { Loader } from '@entur/loader'
-import { toggleValueInList, isDarkOrDefaultTheme } from '../../../../utils'
+import { isDarkOrDefaultTheme } from '../../../../utils'
+import { toggleValueInList } from '../../../../utils/array'
 import { StopPlaceWithLines } from '../../../../types'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { ThemeContrastWrapper } from '../../../ThemeWrapper/ThemeContrastWrapper'
