+++ src/logic/useStopPlacesWithLines.ts
@@ -1,7 +1,8 @@
 import { useEffect, useState } from 'react'
 import { Line, StopPlaceWithLines } from '../types'
-import { createAbortController, unique } from '../utils'
+import { createAbortController } from '../utils'
 import { useSettings } from '../settings/SettingsProvider'
+import { unique } from '../utils/array'
 import { getStopPlacesWithLines } from './get-stop-places-with-lines/getStopPlacesWithLines'
 import { useStopPlacesWithDepartures } from './use-stop-places-with-departures/useStopPlacesWithDepartures'
